Settings editted

+++ src/Data.java
@@ -20,7 +20,7 @@ public class Data {
 	public static int totalQuestions =0; 
 	public static PlayerAccountList accounts = new PlayerAccountList();
 	public static String userName;
-	public static int seconds;
+	public static double seconds;
 
 	public static int getPoints () {
 		return points;
